+++ crates/vmi-os-windows/src/lib.rs
@@ -62,6 +62,7 @@ use zerocopy::{FromBytes, IntoBytes};
 
 mod arch;
 use self::arch::ArchAdapter;
+pub use self::arch::{WindowsExceptionVector, WindowsInterrupt, WindowsPageTableEntry};
 
 mod error;
 pub use self::error::WindowsError;